third_party/blink/web_tests/accessibility/description-calc-ariaDescribedByElements.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
function setAriaDescribedByElements(id) {
    const el = document.getElementById(id);
    if (!el) {
	throw new Error(`no element with id "${id}"`);
    }

    const describedByAttribute = el.getAttribute("data-describedby");
    const describedByIds = describedByAttribute.split(/\s+/);
    const describedByElements = describedByIds.map((id) => {
        return document.getElementById(id);
    });
    el.ariaDescribedByElements = describedByElements;
}
</script>

<div class="container">
    <button id="describedby" data-describedby="description1">Contents</button>
    <div id="description1">Description 1</div>
</div>

<script>
test(function(t) {
    setAriaDescribedByElements("describedby");

    var axDescribedby = accessibilityController.accessibleElementById("describedby");
    assert_equals(axDescribedby.name, "Contents");
    assert_equals(axDescribedby.nameFrom, "contents");
    assert_equals(axDescribedby.description, "Description 1");
    assert_equals(axDescribedby.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby.descriptionElementCount(), 1);
    assert_equals(axDescribedby.descriptionElementAtIndex(0).role, "AXRole: AXGenericContainer");
}, "Button with ariaDescribedByElements referencing the id of another element.");

test(function(t){
    setAriaDescribedByElements("describedby");

    var describedby = document.getElementById("describedby");
    describedby.ariaDescribedByElements = null;

    var axDescribedby = accessibilityController.accessibleElementById("describedby");
    assert_equals(axDescribedby.name, "Contents");
    assert_equals(axDescribedby.nameFrom, "contents");
    assert_equals(axDescribedby.description, "");
    assert_equals(axDescribedby.descriptionFrom, "");
    assert_equals(axDescribedby.descriptionElementCount(), 0);
}, "Setting ariaDescribedByElements to null removes the description.");

</script>

<div class="container">
    <button id="describedby2" data-describedby="describedby2">Contents</button>
</div>

<script>
test(function(t){
    setAriaDescribedByElements("describedby2");

    var axDescribedby2 = accessibilityController.accessibleElementById("describedby2");
    assert_equals(axDescribedby2.name, "Contents");
    assert_equals(axDescribedby2.nameFrom, "contents");
    assert_equals(axDescribedby2.description, "Contents");
    assert_equals(axDescribedby2.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby2.descriptionElementCount(), 1);
    assert_equals(axDescribedby2.descriptionElementAtIndex(0).role, "AXRole: AXButton");
}, "Button with ariaDescribedByElements referencing itself.");
</script>

<div class="container">
    <button id="describedby3" data-describedby="describedby3 description3">Contents</button>
    <div id="description3">Description 3</div>
</div>

<script>
test(function(t){
    setAriaDescribedByElements("describedby3");

    var axDescribedby3 = accessibilityController.accessibleElementById("describedby3");
    assert_equals(axDescribedby3.name, "Contents");
    assert_equals(axDescribedby3.nameFrom, "contents");
    assert_equals(axDescribedby3.description, "Contents Description 3");
    assert_equals(axDescribedby3.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby3.descriptionElementCount(), 2);
    assert_equals(axDescribedby3.descriptionElementAtIndex(0).role, "AXRole: AXButton");
    assert_equals(axDescribedby3.descriptionElementAtIndex(1).role, "AXRole: AXGenericContainer");
}, "Button with ariaDescribedByElements referencing itself and the id of another element.");
</script>

<div class="container">
    <button id="describedby4">Contents</button>
</div>

<script>
test(function(t){
    var describedby4 = document.getElementById("describedby4");
    describedby4.ariaDescribedByElements = [];

    var axDescribedby4 = accessibilityController.accessibleElementById("describedby4");
    assert_equals(axDescribedby4.name, "Contents");
    assert_equals(axDescribedby4.nameFrom, "contents");
    assert_equals(axDescribedby4.description, "");
    assert_equals(axDescribedby4.descriptionFrom, "");
    assert_equals(axDescribedby4.descriptionElementCount(), 0);
}, "ariaDescribedByElements reference to empty array.");
</script>

<div class="container">
    <button id="describedby5" data-describedby="description5">Contents</button>
    <div id="description5"></div>
</div>

<script>
test(function(t){
    setAriaDescribedByElements("describedby5");

    var axDescribedby5 = accessibilityController.accessibleElementById("describedby5");
    assert_equals(axDescribedby5.name, "Contents");
    assert_equals(axDescribedby5.nameFrom, "contents");
    assert_equals(axDescribedby5.description, "");
    assert_equals(axDescribedby5.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby5.descriptionElementCount(), 1);
    assert_equals(axDescribedby5.descriptionElementAtIndex(0).role, "AXRole: AXGenericContainer");
}, "ariaDescribedByElements reference to element with no text.");
</script>

<div class="container">
    <button id="describedby6" data-describedby="description6">Contents</button>
    <h3 id="description6" style="visibility: hidden">Invisible description</h3>
</div>

<script>
test(function(t){
    setAriaDescribedByElements("describedby6");

    assert_true(accessibilityController.accessibleElementById("description6").isIgnored);
    var axDescribedby6 = accessibilityController.accessibleElementById("describedby6");
    assert_equals(axDescribedby6.name, "Contents");
    assert_equals(axDescribedby6.nameFrom, "contents");
    assert_equals(axDescribedby6.description, "Invisible description");
    assert_equals(axDescribedby6.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby6.descriptionElementCount(), 1);
    assert_equals(axDescribedby6.descriptionElementAtIndex(0).role, "AXRole: AXHeading");
}, "Button with ariaDescribedByElements reference to invisible element.");
</script>

<div class="container">
    <button id="describedby7" data-describedby="description7">Contents</button>
    <h3 id="description7" style="display: none">Display-none description</h3>
</div>

<script>
test(function(t){
    setAriaDescribedByElements("describedby7");

    assert_true(accessibilityController.accessibleElementById("description7").isIgnored);
var axDescribedby7 = accessibilityController.accessibleElementById("describedby7");
    assert_equals(axDescribedby7.name, "Contents");
    assert_equals(axDescribedby7.nameFrom, "contents");
    assert_equals(axDescribedby7.description, "Display-none description");
    assert_equals(axDescribedby7.descriptionFrom, "relatedElement");
    assert_equals(axDescribedby7.descriptionElementCount(), 1);
    assert_equals(axDescribedby7.descriptionElementAtIndex(0).role, "AXRole: AXHeading");
}, "Button with ariaDescribedByElements reference to display-none element.");
</script>
